project/multi_tool_agent/requirements.txt
google-adk==1.18.0
litellm==1.79.1
fastmcp==2.13.0.2
pyjwt==2.10.1
